changed error to zmodem_error
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@841 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/rzsz/zm.c
@@ -294,7 +294,7 @@ zsendline_s(const char *s, size_t count)
 			t++;
 		}
 		if (t!=s) {
-			raw_write(0,s,(size_t)(t-s));
+			raw_write(0,s,(t-s));
 			lastsent=t[-1];
 			s=t;
 		}
